converted kernels for images to floats rather than ints... slower but I don't want to figure out separability right now

+++ library/imageAnalysis.h
@@ -43,9 +43,9 @@ void equalizeHistogram(int *original, int *mappings, int numPixels);
 
 void equalizeImageWithHist(Image *image, Image *d_equalizedImage, int *h_mappings);
 
-void linearFilter(Image *image, Image *output, int *kernel, int kWidth, int kHeight);
+void linearFilter(Image *image, Image *output, float *kernel, int kWidth, int kHeight);
 
-void medianFilter(Image *image, Image *output, int *kernel, int kWidth, int kHeight);
+void medianFilter(Image *image, Image *output, float *kernel, int kWidth, int kHeight);
 
 void saltAndPepperNoise(Image *image, Image *output, int level);
